+++ crates/swc/tests/tsc-references/emitDefaultParametersFunctionExpressionES6.2.minified.js
@@ -1,2 +1,2 @@
 //// [emitDefaultParametersFunctionExpressionES6.ts]
-!function(num, boo = !1, ...rest) {}(10);
+!function(num, boo = !1, ...rest) {}(0);