+++ crates/swc_ecma_minifier/src/compress/optimize/iife.rs
@@ -157,21 +157,6 @@ where
             Callee::Expr(e) => &mut **e,
         };
 
-        fn find_params(callee: &mut Expr) -> Option<Vec<&mut Pat>> {
-            match callee {
-                Expr::Arrow(callee) => Some(callee.params.iter_mut().collect()),
-                Expr::Fn(callee) => Some(
-                    callee
-                        .function
-                        .params
-                        .iter_mut()
-                        .map(|param| &mut param.pat)
-                        .collect(),
-                ),
-                _ => None,
-            }
-        }
-
         fn clean_params(callee: &mut Expr) {
             match callee {
                 Expr::Arrow(callee) => callee.params.retain(|p| !p.is_invalid()),
@@ -192,17 +177,6 @@ where
                 .is_some()
             {
                 return;
-            }
-        }
-
-        fn find_body(callee: &mut Expr) -> Option<Either<&mut BlockStmt, &mut Expr>> {
-            match callee {
-                Expr::Arrow(e) => match &mut e.body {
-                    BlockStmtOrExpr::BlockStmt(b) => Some(Either::Left(b)),
-                    BlockStmtOrExpr::Expr(b) => Some(Either::Right(&mut **b)),
-                },
-                Expr::Fn(e) => Some(Either::Left(e.function.body.as_mut().unwrap())),
-                _ => None,
             }
         }
 
@@ -315,11 +289,104 @@ where
                     trace_op!("inline: Inlining arguments");
                     optimizer.inline_vars_in_node(body, vars);
                 }
-                _ => {}
+                _ => {
+                    unreachable!("find_body and find_params should match")
+                }
             }
         }
 
         clean_params(callee);
+    }
+
+    /// If a parameter is not used, we can ignore return value of the
+    /// corresponding argument.
+    pub(super) fn ignore_unused_args_of_iife(&mut self, e: &mut CallExpr) {
+        if !self.options.unused {
+            return;
+        }
+
+        let callee = match &mut e.callee {
+            Callee::Super(_) | Callee::Import(_) => return,
+            Callee::Expr(e) => &mut **e,
+        };
+
+        match find_body(callee) {
+            Some(body) => match body {
+                Either::Left(body) => {
+                    if contains_arguments(body) {
+                        return;
+                    }
+                }
+                Either::Right(body) => {
+                    if contains_arguments(body) {
+                        return;
+                    }
+                }
+            },
+            None => return,
+        }
+
+        if let Expr::Fn(FnExpr {
+            ident: Some(ident), ..
+        }) = callee
+        {
+            if self
+                .data
+                .vars
+                .get(&ident.to_id())
+                .filter(|usage| usage.used_recursively)
+                .is_some()
+            {
+                return;
+            }
+        }
+
+        let mut removed = vec![];
+        let params = find_params(callee);
+        if let Some(mut params) = params {
+            // We check for parameter and argument
+            for (idx, param) in params.iter_mut().enumerate() {
+                if let Pat::Ident(param) = &mut **param {
+                    if let Some(usage) = self.data.vars.get(&param.to_id()) {
+                        if usage.ref_count == 0 {
+                            removed.push(idx);
+                        }
+                    }
+                }
+            }
+
+            if removed.is_empty() {
+                log_abort!("`removed` is empty");
+                return;
+            }
+        } else {
+            unreachable!("find_body and find_params should match")
+        }
+
+        for idx in removed {
+            if let Some(arg) = e.args.get_mut(idx) {
+                if arg.spread.is_some() {
+                    break;
+                }
+
+                // Optimize
+                let new = self.ignore_return_value(&mut arg.expr);
+
+                if let Some(new) = new {
+                    arg.expr = Box::new(new);
+                } else {
+                    // Use `0` if it's removed.
+                    arg.expr = Number {
+                        span: arg.expr.span(),
+                        value: 0.0,
+                        raw: None,
+                    }
+                    .into();
+                }
+            } else {
+                break;
+            }
+        }
     }
 
     #[cfg_attr(feature = "debug", tracing::instrument(skip_all))]
@@ -1012,5 +1079,30 @@ where
 
             _ => false,
         }
+    }
+}
+
+fn find_params(callee: &mut Expr) -> Option<Vec<&mut Pat>> {
+    match callee {
+        Expr::Arrow(callee) => Some(callee.params.iter_mut().collect()),
+        Expr::Fn(callee) => Some(
+            callee
+                .function
+                .params
+                .iter_mut()
+                .map(|param| &mut param.pat)
+                .collect(),
+        ),
+        _ => None,
+    }
+}
+fn find_body(callee: &mut Expr) -> Option<Either<&mut BlockStmt, &mut Expr>> {
+    match callee {
+        Expr::Arrow(e) => match &mut e.body {
+            BlockStmtOrExpr::BlockStmt(b) => Some(Either::Left(b)),
+            BlockStmtOrExpr::Expr(b) => Some(Either::Right(&mut **b)),
+        },
+        Expr::Fn(e) => Some(Either::Left(e.function.body.as_mut().unwrap())),
+        _ => None,
     }
 }